Bump debian version number
So this was in an attempt to get the i386 build to work on Launchpad.
The attempt did not succeed; stuff still breaks with what appears to be
multilib errors. I don't believe the launchpad is building multilib
stuff wrong, so something must be interacting weird between this
package, its upstream dependencies, and the multilib build system.
*However* since this stuff is for Travis, which runs amd64, I'm just
going to ignore the problem for now.